requirements.txt
matplotlib>=3.1.0
numpy>=1.16.4
ruamel.yaml>=0.16.5
ruamel.yaml.clib>=0.1.2
qpsolvers>=1.0.7
commonroad-io>=2020.2
